+++ packages/core-files/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'unchained:core-files',
-  version: '1.0.0-beta5',
+  version: '1.0.0-beta7',
   summary: 'Unchained Engine Core: Files',
   git: 'https://github.com/unchainedshop/unchained',
   documentation: 'README.md',
@@ -12,9 +12,9 @@ Npm.depends({
 
 Package.onUse((api) => {
   api.versionsFrom('2.2');
-  api.use('ostrio:files@1.14.3');
+  api.use('ostrio:files@2.0.1');
   api.use('ecmascript');
-  api.use('unchained:core-settings@1.0.0-beta5');
+  api.use('unchained:core-settings@1.0.0-beta7');
   api.mainModule('core-files.js');
 });