+++ packages/core-messaging/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'unchained:core-messaging',
-  version: '1.0.0-beta5',
+  version: '1.0.0-beta7',
   summary: 'Unchained Engine Core: Messaging',
   git: 'https://github.com/unchainedshop/unchained',
   documentation: 'README.md',
@@ -14,8 +14,8 @@ Npm.depends({
 Package.onUse((api) => {
   api.versionsFrom('2.2');
   api.use('ecmascript');
-  api.use('unchained:core-logger@1.0.0-beta5');
-  api.use('unchained:core-worker@1.0.0-beta5');
+  api.use('unchained:core-logger@1.0.0-beta7');
+  api.use('unchained:core-worker@1.0.0-beta7');
 
   api.mainModule('messaging.js', 'server');
 });