+++ packages/platform/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'unchained:platform',
-  version: '1.0.0-beta5',
+  version: '1.0.0-beta7',
   summary: 'Unchained Engine',
   git: 'https://github.com/unchainedshop/unchained',
   documentation: 'README.md',
@@ -20,8 +20,8 @@ Package.onUse((api) => {
   api.use('email');
   api.use('percolate:migrations@1.0.2');
 
-  api.use('unchained:core@1.0.0-beta5');
-  api.use('unchained:api@1.0.0-beta5');
+  api.use('unchained:core@1.0.0-beta7');
+  api.use('unchained:api@1.0.0-beta7');
 
   api.mainModule('platform.js', 'server');
 });