+++ tests/assortment-filter.test.js
@@ -283,7 +283,7 @@ describe('AssortmentFilter', () => {
   describe('mutation.removeAssortmentFilter for admin users should', () => {
     it('remove assortment filter successfuly when passed valid ID', async () => {
       const {
-        // eslint-disable-next-line no-unused-vars
+        // eslint-disable-next-line
         data: { removeAssortmentFilter },
       } = await graphqlFetch({
         query: /* GraphQL */ `